% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.ach.R
\name{get.ach}
\alias{get.ach}
\title{Get achievement data}
\usage{
get.ach(study = NULL, year = NULL, specification = NULL)
}
\arguments{
\item{study}{a character value indicating the ILSA name.}

\item{year}{a numeric value indicating the year of the ILSA cycle.}

\item{specification}{a character value indicating extra specification like grade
(e.g., \code{"G8"} for TIMSS) or subject (e.g., \code{"Math"} for TIMSSADVANCED).}
}
\description{
Get achievement data
}
\examples{

get.ach("timss",1995,"g4")


}
